refactor(providers-section): extract PromptManager + ProvidersList sub-component
Was 710 lines: AI provider CRUD + title-generation prompt management
sharing a tab bar but otherwise independent (separate fetches, separate
dialogs, separate state). Split into:

- prompt-manager.tsx -- owns selectedPromptId / customPrompts state,
  the GET /api/settings/user fetch + PUT round-trip, both dialogs
  (view + edit/create), and all rendering. Wired to useSettings()
  exactly as before for the loading/saving banner state.
- providers-section.tsx (286 lines) -- AI providers list + tab bar,
  delegates the prompts tab entirely to <PromptManager />. The
  configured-providers list rendering also moves out to a local
  <ProvidersList /> sub-component so the parent's render is just
  header + tabs + active-pane.

The two halves were 100% logically independent already; the line-count
win is real (parent file effectively cut in half) and the surface area
each component has to reason about shrinks even more (PromptManager
owns 7 of the original 10 useState calls).

No behavior change.
